# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: zerollama API
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Live OpenAPI for this zerollama server.

    Includes the native Ollama `/api/*` surface plus OpenAI-compatible
    `/v1/*` routes (chat, embeddings, images, audio speech/transcriptions,
    voice catalog, and video generation).

    LocalAI control plane (LA7/LA9/LA11/LA11b/LA15/LA17): `/api/repair`,
    `/api/score`, `/api/router/*`, `/api/aliases`, experimental web_search/fetch.
    In-band alias and router rewrite may set `X-Zerollama-*` response headers.
    Failed-load cooldown (LA18) returns 503 with `Retry-After`.

    Interactive docs: `GET /docs` · raw: `GET /openapi.yaml` · `GET /openapi.json`
servers:
  - url: http://localhost:11434
    description: This server (rewritten at serve time to the request host)
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
...
  schemas:
...
    InferenceError:
      type: object
      description: >
        Error body on generate/chat (JSON or NDJSON). Busy 503s include retry_after;
        stream/runner failures may include cause=host_unstable and partial timings;
        pin/VRAM conflicts use cause=runtime_pin_gguf or runtime_pin_ggml.
        LA18 failed-load cooldown also returns 503 + Retry-After (error mentions cooldown).
        M15e: QoS wait-abort / busy may include preempted_reason; per-call timeout
        uses HTTP 504 with timeout_seconds (not this schema alone).
      required: [error]
      properties:
        error:
          type: string
        status:
          type: integer
        retry_after:
          type: integer
          description: Seconds to wait before retry (also sent as Retry-After header on 503)
        preempted_reason:
          type: string
          description: >
            Why the request aborted while deferred behind higher QoS (e.g.
            lower_wait_interactive, wait_parent, fulfillment_exclusive). Why:
            Hermes needs to distinguish QoS wait abort from generic cancel /
            queue-full. Not mid-stream hard preemption.
        timeout_seconds:
          type: number
          description: >
            Present on HTTP 504 when per-call `timeout` fired (DeadlineExceeded).
            Distinct from 499 client cancel.
        cause:
          type: string
          description: >
            host_unstable (runner/process exit); runtime_pin_gguf (active pin holds a
            different Python GGUF); runtime_pin_ggml (pinned/in-use ggml still resident
            after pin-respecting unload — resume blocked to avoid dual-stack OOM);
            runtime_vram (other prepareRuntimeVRAM failures).
          enum:
            - host_unstable
            - runtime_pin_gguf
            - runtime_pin_ggml
            - runtime_vram
        total_duration:
          type: integer
          description: Nanoseconds wall time until failure
        load_duration:
          type: integer
          description: Nanoseconds spent loading before failure
        time_to_first_token:
          type: integer
          description: Nanoseconds to first token when a token was emitted
...
    CanLoadRequest:
      type: object
      required: [model]
      properties:
        model:
          type: string
        options:
          type: object
          additionalProperties: true
          description: Optional num_ctx, gguf, etc.
    CanLoadResponse:
      type: object
      description: >
        Always HTTP 200. confidence=exact uses runtime VRAM estimate; heuristic
        is ggml count/concurrency only (does not guarantee VRAM fit).
      properties:
        model:
          type: string
        backend:
          type: string
        can_load:
          type: boolean
        confidence:
          type: string
          enum: [exact, heuristic]
        already_loaded:
          type: boolean
        needs_eviction:
          type: boolean
        eviction_reason:
          type: string
        busy:
          type: boolean
        busy_reason:
          type: string
        loads_paused:
          type: boolean
        queue:
          type: object
          additionalProperties: true
        warm:
          type: object
          additionalProperties: true
        vram_estimate:
          type: object
          additionalProperties: true
        vram_budget:
          type: object
          additionalProperties: true
        suggested_max_num_ctx:
          type: integer
        max_loaded_models:
          type: integer
        loaded_count:
          type: integer
        device_count:
          type: integer
          description: >
            Host GPU count (runtime topology or ggml discover). Why: Hermes multi-GPU
            awareness without scraping /health. Not a TP planner.
        tensor_parallel:
          type: integer
          description: Effective TP width from runtime config (often 1 on single GPU)
        split_mode:
          type: string
          description: Runtime split_mode (e.g. layer, row, tensor) when TP > 1
        tensor_split:
          type: array
          items:
            type: number
          description: Per-device split weights when configured
        main_gpu:
          type: integer
          description: Primary GPU index from runtime config
        notes:
          type: string
...
paths:
...
  /api/can-load:
    post:
      summary: Capacity dry-run (admit without loading)
      description: >
        Returns whether a model can be admitted under known checks without calling
        GetRunner. Always HTTP 200 with structured fields. Runtime path uses
        VRAM estimate (confidence=exact); ggml is count/group heuristic only.
        Fail closed when estimate/GGUF is unavailable. already_loaded requires
        GGUF path match (not merely any loaded model). needs_eviction means a
        swap would occur — thrash-sensitive clients should require !needs_eviction.
        M15e: surfaces host topology (device_count, tensor_parallel, split_mode,
        tensor_split, main_gpu) from runtime estimate — not a TP planner.
      operationId: canLoad
      x-codeSamples:
        - lang: bash
          label: Default
          source: |
            curl http://localhost:11434/api/can-load -d '{"model":"llama3.2:latest","options":{"num_ctx":8192}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanLoadRequest"
      responses:
        "200":
          description: Dry-run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanLoadResponse"
        "400":
          description: Missing model or invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InferenceError"
...